Test blockBroadcaster
