+++ packages/p2p/source/network-monitor.test.ts
@@ -1,5 +1,5 @@
 import { Identifiers } from "@mainsail/contracts";
-import { Enums, Providers, Utils } from "@mainsail/kernel";
+import { Enums, Providers } from "@mainsail/kernel";
 import { BigNumber } from "@mainsail/utils";
 import importFresh from "import-fresh";
 
@@ -1127,299 +1127,5 @@ describeSkip<{
 				fromBlockHeight: match.number,
 			});
 		}
-	});
-
-	each(
-		"#broadcastBlock - should not broadcast to any peer when blockPing >= 4",
-		async ({ context, dataset }) => {
-			const count = dataset;
-
-			const block = {
-				data: {
-					blockSignature:
-						"3045022100e7385c6ea42bd950f7f6ab8c8619cf2f66a41d8f8f185b0bc99af032cb25f30d02200b6210176a6cedfdcbe483167fd91c21d740e0e4011d24d679c601fdd46b0de9",
-					generatorPublicKey: "026c598170201caf0357f202ff14f365a3b09322071e347873869f58d776bfc565",
-					height: 2,
-					id: "17882607875259085966",
-					numberOfTransactions: 0,
-					payloadHash: "e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855",
-					payloadLength: 0,
-					previousBlock: "17184958558311101492",
-					reward: BigNumber.make("0"),
-					timestamp: 46_583_330,
-					totalAmount: BigNumber.make("0"),
-					totalFee: BigNumber.make("0"),
-					version: 0,
-				},
-				transactions: [],
-			};
-
-			const peers = [
-				new Peer("180.177.54.4", 4000),
-				new Peer("181.177.54.4", 4000),
-				new Peer("182.177.54.4", 4000),
-				new Peer("183.177.54.4", 4000),
-				new Peer("184.177.54.4", 4000),
-			];
-			stub(repository, "getPeers").returnValue(peers);
-
-			stub(blockchain, "getBlockPing").returnValueOnce({
-				block: block.data,
-				count,
-				first: 10_200,
-				last: 10_900,
-			});
-			const spyCommunicatorPostBlock = spy(communicator, "postBlock");
-
-			await context.networkMonitor.broadcastBlock(block);
-
-			spyCommunicatorPostBlock.neverCalled();
-		},
-		[4, 5, 10, 50],
-	);
-
-	each(
-		"#broadcastBlock - should broadcast to (4 - blockPing)/4 of our peers when blockPing < 4",
-		async ({ context, dataset }) => {
-			const count = dataset;
-
-			const block = {
-				data: {
-					blockSignature:
-						"3045022100e7385c6ea42bd950f7f6ab8c8619cf2f66a41d8f8f185b0bc99af032cb25f30d02200b6210176a6cedfdcbe483167fd91c21d740e0e4011d24d679c601fdd46b0de9",
-					generatorPublicKey: "026c598170201caf0357f202ff14f365a3b09322071e347873869f58d776bfc565",
-					height: 2,
-					id: "17882607875259085966",
-					numberOfTransactions: 0,
-					payloadHash: "e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855",
-					payloadLength: 0,
-					previousBlock: "17184958558311101492",
-					reward: BigNumber.make("0"),
-					timestamp: 46_583_330,
-					totalAmount: BigNumber.make("0"),
-					totalFee: BigNumber.make("0"),
-					version: 0,
-				},
-				transactions: [],
-			};
-
-			const peers = [
-				new Peer("180.177.54.4", 4000),
-				new Peer("181.177.54.4", 4000),
-				new Peer("182.177.54.4", 4000),
-				new Peer("183.177.54.4", 4000),
-				new Peer("184.177.54.4", 4000),
-			];
-			stub(repository, "getPeers").returnValue(peers);
-
-			stub(blockchain, "getBlockPing").returnValueOnce({
-				block: block.data,
-				count,
-				first: 10_200,
-				last: 10_900,
-			});
-			const spyCommunicatorPostBlock = spy(communicator, "postBlock");
-
-			await context.networkMonitor.broadcastBlock(block);
-
-			spyCommunicatorPostBlock.calledTimes(Math.ceil((peers.length * (4 - count)) / 4));
-		},
-		[0, 1, 2, 3],
-	);
-
-	each(
-		"#broadcastBlock - should broadcast to all of our peers when block.id doesnt match blockPing.id",
-		async ({ context, dataset }) => {
-			const count = dataset;
-
-			const block = {
-				data: {
-					blockSignature:
-						"3045022100e7385c6ea42bd950f7f6ab8c8619cf2f66a41d8f8f185b0bc99af032cb25f30d02200b6210176a6cedfdcbe483167fd91c21d740e0e4011d24d679c601fdd46b0de9",
-					generatorPublicKey: "026c598170201caf0357f202ff14f365a3b09322071e347873869f58d776bfc565",
-					height: 2,
-					id: "17882607875259085966",
-					numberOfTransactions: 0,
-					payloadHash: "e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855",
-					payloadLength: 0,
-					previousBlock: "17184958558311101492",
-					reward: BigNumber.make("0"),
-					timestamp: 46_583_330,
-					totalAmount: BigNumber.make("0"),
-					totalFee: BigNumber.make("0"),
-					version: 0,
-				},
-				transactions: [],
-			};
-
-			const temporaryBlock = Utils.cloneDeep(block);
-
-			temporaryBlock.data.id = "random_id";
-
-			const peers = [
-				new Peer("180.177.54.4", 4000),
-				new Peer("181.177.54.4", 4000),
-				new Peer("182.177.54.4", 4000),
-				new Peer("183.177.54.4", 4000),
-				new Peer("184.177.54.4", 4000),
-			];
-			stub(repository, "getPeers").returnValue(peers);
-
-			stub(blockchain, "getBlockPing").returnValueOnce({
-				block: temporaryBlock.data,
-				count,
-				first: 10_200,
-				last: 10_900,
-			});
-			const spyCommunicatorPostBlock = spy(communicator, "postBlock");
-
-			await context.networkMonitor.broadcastBlock(block);
-
-			spyCommunicatorPostBlock.calledTimes(peers.length);
-		},
-		[0, 1, 2, 3],
-	);
-
-	it("#broadcastBlock - should not wait if block is from forger, when blockPing.last - blockPing.first < 500ms", async ({
-		networkMonitor,
-	}) => {
-		const block = {
-			data: {
-				blockSignature:
-					"3045022100e7385c6ea42bd950f7f6ab8c8619cf2f66a41d8f8f185b0bc99af032cb25f30d02200b6210176a6cedfdcbe483167fd91c21d740e0e4011d24d679c601fdd46b0de9",
-				generatorPublicKey: "026c598170201caf0357f202ff14f365a3b09322071e347873869f58d776bfc565",
-				height: 2,
-				id: "17882607875259085966",
-				numberOfTransactions: 0,
-				payloadHash: "e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855",
-				payloadLength: 0,
-				previousBlock: "17184958558311101492",
-				reward: BigNumber.make("0"),
-				timestamp: 46_583_330,
-				totalAmount: BigNumber.make("0"),
-				totalFee: BigNumber.make("0"),
-				version: 0,
-			},
-			transactions: [],
-		};
-
-		const peers = [
-			new Peer("180.177.54.4", 4000),
-			new Peer("181.177.54.4", 4000),
-			new Peer("182.177.54.4", 4000),
-			new Peer("183.177.54.4", 4000),
-			new Peer("184.177.54.4", 4000),
-		];
-		stub(repository, "getPeers").returnValue(peers);
-
-		stub(blockchain, "getBlockPing").returnValue({
-			block: block.data,
-			count: 2,
-			first: 10_200,
-			fromForger: true,
-			last: 10_500,
-		});
-		const spyCommunicatorPostBlock = spy(communicator, "postBlock");
-		// const spySleep = stub(Utils, "sleep"); // TODO: Test sleep
-
-		await networkMonitor.broadcastBlock(block);
-
-		spyCommunicatorPostBlock.calledTimes(peers.length);
-		// spySleep.neverCalled();
-	});
-
-	it("#broadcastBlock - should wait until 500ms have elapsed between blockPing.last and blockPing.first before broadcasting, when blockPing.last - blockPing.first < 500ms", async ({
-		networkMonitor,
-	}) => {
-		const block = {
-			data: {
-				blockSignature:
-					"3045022100e7385c6ea42bd950f7f6ab8c8619cf2f66a41d8f8f185b0bc99af032cb25f30d02200b6210176a6cedfdcbe483167fd91c21d740e0e4011d24d679c601fdd46b0de9",
-				generatorPublicKey: "026c598170201caf0357f202ff14f365a3b09322071e347873869f58d776bfc565",
-				height: 2,
-				id: "17882607875259085966",
-				numberOfTransactions: 0,
-				payloadHash: "e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855",
-				payloadLength: 0,
-				previousBlock: "17184958558311101492",
-				reward: BigNumber.make("0"),
-				timestamp: 46_583_330,
-				totalAmount: BigNumber.make("0"),
-				totalFee: BigNumber.make("0"),
-				version: 0,
-			},
-			transactions: [],
-		};
-
-		const peers = [
-			new Peer("180.177.54.4", 4000),
-			new Peer("181.177.54.4", 4000),
-			new Peer("182.177.54.4", 4000),
-			new Peer("183.177.54.4", 4000),
-			new Peer("184.177.54.4", 4000),
-		];
-		stub(repository, "getPeers").returnValue(peers);
-
-		stub(blockchain, "getBlockPing").returnValue({ block: block.data, count: 2, first: 10_200, last: 10_500 });
-		const spyCommunicatorPostBlock = spy(communicator, "postBlock");
-		// const spySleep = jest.spyOn(Utils, "sleep"); // TODO: Test sleep
-
-		await networkMonitor.broadcastBlock(block);
-
-		spyCommunicatorPostBlock.calledTimes(Math.ceil((peers.length * 1) / 2));
-
-		// expect(spySleep).toBeCalledTimes(1);
-		// expect(spySleep).toBeCalledWith(200); // 500 - (last - first)
-	});
-
-	it("#broadcastBlock - should not broadcast if during waiting we have received a new block, when blockPing.last - blockPing.first < 500ms", async ({
-		networkMonitor,
-	}) => {
-		const block = {
-			data: {
-				blockSignature:
-					"3045022100e7385c6ea42bd950f7f6ab8c8619cf2f66a41d8f8f185b0bc99af032cb25f30d02200b6210176a6cedfdcbe483167fd91c21d740e0e4011d24d679c601fdd46b0de9",
-				generatorPublicKey: "026c598170201caf0357f202ff14f365a3b09322071e347873869f58d776bfc565",
-				height: 2,
-				id: "17882607875259085966",
-				numberOfTransactions: 0,
-				payloadHash: "e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855",
-				payloadLength: 0,
-				previousBlock: "17184958558311101492",
-				reward: BigNumber.make("0"),
-				timestamp: 46_583_330,
-				totalAmount: BigNumber.make("0"),
-				totalFee: BigNumber.make("0"),
-				version: 0,
-			},
-			transactions: [],
-		};
-
-		const peers = [
-			new Peer("180.177.54.4", 4000),
-			new Peer("181.177.54.4", 4000),
-			new Peer("182.177.54.4", 4000),
-			new Peer("183.177.54.4", 4000),
-			new Peer("184.177.54.4", 4000),
-		];
-		stub(repository, "getPeers").returnValue(peers);
-
-		stub(blockchain, "getBlockPing")
-			.returnValueNth(0, { block: block.data, count: 2, first: 10_200, last: 10_500 })
-			.returnValueNth(1, {
-				block: { ...block.data, height: 3, id: "11111111" },
-				count: 2,
-				first: 10_200,
-				last: 10_500,
-			});
-
-		const spyCommunicatorPostBlock = spy(communicator, "postBlock");
-		// const spySleep = jest.spyOn(Utils, "sleep"); // TODO: Test sleep
-
-		await networkMonitor.broadcastBlock(block);
-
-		spyCommunicatorPostBlock.neverCalled();
-
-		// expect(spySleep).toBeCalledTimes(1);
 	});
 });
